pyproject.toml
[project]
name = "sandipani-vidhyapeeth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.12.36",
]
